implemented AdSearchService and LocalityService

+++ src/main/java/io/github/aglushkovsky/advertisingservice/dto/request/FindAllAdsFilterRequestDto.java
@@ -1,9 +1,11 @@
 package io.github.aglushkovsky.advertisingservice.dto.request;
 
+import java.math.BigDecimal;
+
 public record FindAllAdsFilterRequestDto(String term,
                                          Boolean onlyInTitle,
-                                         Long minPrice,
-                                         Long maxPrice,
+                                         BigDecimal minPrice,
+                                         BigDecimal maxPrice,
                                          Long publisherId,
                                          Long localityId,
                                          Long limit,
